+++ react-components/src/pages/404Page/404Page.tsx
@@ -2,8 +2,6 @@ import { FC } from 'react'
 
 import styles from './404Page.module.scss'
 
-const Page404: FC = () => {
-  return <div className={styles.content}>404 Page</div>
-}
+const Page404: FC = () => <div className={styles.content}>404 Page</div>
 
 export default Page404
