+++ react-components/src/pages/AboutUsPage/AboutUsPage.tsx
@@ -3,14 +3,12 @@ import Layout from '../../components/Layout'
 
 import styles from './AboutUsPage.module.scss'
 
-const AboutUsPage: FC = () => {
-  return (
-    <div>
-      <Layout currentPage="About">
-        <div className={styles.content}>About Us Page</div>
-      </Layout>
-    </div>
-  )
-}
+const AboutUsPage: FC = () => (
+  <div>
+    <Layout currentPage="About">
+      <div className={styles.content}>About Us Page</div>
+    </Layout>
+  </div>
+)
 
 export default AboutUsPage
